Implement Configuration#billable_name_column option
This is used to specify a method to run against the Account's billable association to get a more human-readable name from.
Although it currently works with any method, we will likely have to use it in our queries later on, so the verbiage implies that a table column should be used.

Signed-off-by: Joshua Stowers <5009837+stowersjoshua@users.noreply.github.com>