Add Offerable::Scope class
The Scope class will slowly begin phasing out the existing scopes hash defined on Offerable.
This is needed because some tasks (such as converting options: :auto into an array of queryable values) are too cumbersome to put entirely into view helpers.

We should take our time phasing it in to ensure that it is built to our needs our first time around.

Signed-off-by: Joshua Stowers <5009837+stowersjoshua@users.noreply.github.com>