Implement inventory whitelisting
This will enable users to keep a products's inventory item within scope while still taking it off the market.

Rationale: A basket that has been sold would still be expected to be defined as that same product. Nothing has changed apart from its availability.

This is being added so that subscription packages don't have to disappear after being disabled on one of the applications using the gem.

We will probably want to update the inventory_whitelist method to accept queryable arguments instead. This will do the trick for now though.

Signed-off-by: Joshua Stowers <5009837+stowersjoshua@users.noreply.github.com>