Add Inventory class
This will give us a place to do our price calculations once we have the price_overrides table in place.

The design for this class was based on the Transactions class.

We should ensure that all the scoping rules are specified in instance variables so that we only have to do the query one time (like how ActiveRecord queries work).

Signed-off-by: Joshua Stowers <5009837+stowersjoshua@users.noreply.github.com>