Define price_overrides migrations
This is needed so that we can add a PriceOverride model.

The original migration template file has been updated to include this so that new installations will get a complete migration file.
A price_overrides_migration.rb.tt has been added and will run only for old installations that ran the original version of the migration file. We determine if this is the case by checking if their existing schema includes a definition for the new model's table.

Signed-off-by: Joshua Stowers <5009837+stowersjoshua@users.noreply.github.com>