Add InventoryItem#available? method
This is done by abstracting out the querying logic to its own method, and using it to grab the available records' ids.

The ids are collected off of a separate Inventory object to prevent accidentally pulling ids from a tighter scope.
For example, if this method gets called from some new "unavailable" scope in the future, it would always return an empty array.

The id-pulling is done only when explicitly called (and not during initialization) so that the extra query is only done if needed.

Signed-off-by: Joshua Stowers <5009837+stowersjoshua@users.noreply.github.com>