Check for price override in OrderItem#default_price
`def price_override` returns false if no record is found so that it doesn't slip past the memoization guard and run the query again.

`def source_records` is now memoized as well. It guards against `nil?` instead of `present?` so that the query does not get run again if no records were found.

We should be careful not to cause an N+1 if we try to reference `default_price` in bulk.

Signed-off-by: Joshua Stowers <5009837+stowersjoshua@users.noreply.github.com>