Save order items from OrderGroups#create action
This way we can add an item to a new cart instead of an existing one.

At some point we should consider allowing OrderGroup to accept nested attributes for order_items so that this can be done from a form, too.

I have opted against implementing that for now since the use-case is for a button and not a form, and nobody should be expected to write out such a complicated button for such a simple task.

This implementation does not use the `current_order_group` since it isn't guaranteed to be empty.
As a result, it is possible that we end up with lots of empty user groups - but we already will from all the guest carts anyways.

Signed-off-by: Joshua Stowers <5009837+stowersjoshua@users.noreply.github.com>