Update Payment to optionally belong to BillingConfiguration
This commit allows payment to optionally belong to a
BillingConfiguration. This way users of the main app
that has mounted A12y can destroy billing configurations
that have been used to make payments.

In the future, it may make more sense to de-scope
billing configurations, but we can revisit this once
we implement a second payment provider and see what
the reprucussions of doing this would be.

Signed-off-by: Evan Gray <evanscottgray@gmail.com>