Add TODO: reminder to fix Zip casting
It appears that somewhere between billing_cpnfigurations being built and it
being saved, the value of zip will be type casted into an integer. This is an
issue if the value was alphanumeric, causing it to result as 0 and pass
validations.

A bigger part of this issue is that we do not allow for alphanumeric values.
Perhaps solving for that will also solve this issue.

Signed-off-by: Avery Hankins <Averyhankinstech@gmail.com>